player decide on card values

+++ player/java/src/main/java/com/devillsroom/poker/player/Card.java
@@ -2,9 +2,9 @@ package com.devillsroom.poker.player;
 
 public class Card {
     final public int suite;
-    final public short value;
+    final public int value;
 
-    public Card(int s, short v) {
+    public Card(int s, int v) {
         suite = s;
         value = v;
     }
