+++ player/java/src/main/java/com/devillsroom/poker/player/Player.java
@@ -4,5 +4,5 @@ public interface Player {
 
     long doBet(long pot, long to_call, long minimum_raise);
 
-    void addCard(int suite, short value);
+    void addCard(int suite, int value);
 }
