+++ player/java/src/test/java/com/devillsroom/poker/player/RaiseOnPairOrCheckPlayerTest.java
@@ -7,17 +7,17 @@ public class RaiseOnPairOrCheckPlayerTest {
     @Test
     public void hasNoPair() {
         RaiseOnPairOrCheckPlayer p = new RaiseOnPairOrCheckPlayer();
-        p.addCard(1, (short)1);
-        p.addCard(1, (short)2);
-        p.addCard(1, (short)3);
+        p.addCard(1, 1);
+        p.addCard(1, 2);
+        p.addCard(1, 3);
         Assert.assertFalse(p.hasPair());
     }
     @Test
     public void hasPair() {
         RaiseOnPairOrCheckPlayer p = new RaiseOnPairOrCheckPlayer();
-        p.addCard(1, (short)1);
-        p.addCard(1, (short)2);
-        p.addCard(1, (short)1);
+        p.addCard(1, 1);
+        p.addCard(1, 2);
+        p.addCard(1, 1);
         Assert.assertTrue(p.hasPair());
     }
 }
